shodanIA Github/requirements.txt
Django>=5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.4
shodan>=1.31.0
python-dotenv>=1.0
openai>=1.40.0
